add more jobserver message types
master cannot blindly send its tokens to the socket buffers and
decrement its token count; otherwise it can leak tokens if nothing ever
reads the token from the socket, so we need slaves to make token
requests before master writes them to the socket.

but that's not enough either, since a slave may acquire a local token
after requesting a remote token from the master, but before reading it
from the socket, resulting in a similar leak. so the slaves need to ack
offered tokens, and the master can decrement its token count only when
it receives an ack.

the master keeps track of how many token requests it has received. to
translate this into the number of jobs waiting for a token, we need to
keep track of offers sent to the socket that were subsequently not
acked. when a slave that has made a request (but not yet received an
offer) uses a local token, it sends a cancel message to master so that
master can keep accurate counters.

it's possible for the master to have sent an offer that has not been
acked when it receives a cancel message. this is fine; the offer remains
in the socket buffers until some slave reads it from there. the number
of offers currently in the socket cannot exceed the number of available
tokens on master, or the number of jobs waiting for a token.