never unset MAKEJOBSEMAPHORE
since -x (and -j) stay in MAKEFLAGS, new child makes could start their
own jobservers if MAKEJOBSEMAPHORE isn't set. so make sure not to unset
it even if we fail to initialize the slave; new children will try to
initialize themselves and properly fall back to sequential if they fail
too.

an alternate approach could be to remove -x and -j from MAKEFLAGS when
we jobserver_disable(), but we don't want to do that (for now).